requirements.txt
python-telegram-bot
pandas
numpy
pyhocon
plotly
nbformat>=4.2.0
kaleido==0.1.0
fpdf2
lxml
